[TASK] move memo tests to payment-test as post payment tests
