+++ test/fixtures/index.js
@@ -7,4 +7,3 @@ module.exports.payments = require('./payments');
 module.exports.settings = require('./settings');
 module.exports.balances = require('./balances');
 module.exports.trustlines = require('./trustlines');
-module.exports.memos = require('./memos');
